requirements.txt
fastapi
uvicorn
pandas
sqlalchemy
psycopg2-binary
sentence-transformers
tqdm
numpy
scikit-learn
python-dotenv
requests
openai
